+++ tests/test_gwas.py
@@ -9,6 +9,8 @@ import subprocess
 import tempfile
 
 
+pth = os.path.join('singularity', 'gwas.sif')
+
 def test_gwas_plink():
     pth = os.path.join('singularity', 'gwas.sif')
     call = f'singularity run {pth} plink --version'
@@ -18,6 +20,16 @@ def test_gwas_plink():
 def test_gwas_plink2():
     pth = os.path.join('singularity', 'gwas.sif')
     call = f'singularity run {pth} plink2 --version'
+    out = subprocess.run(call.split(' '))
+    assert out.returncode == 0
+
+def test_gwas_minimac4():
+    call = f'singularity run {pth} minimac4 --version'
+    out = subprocess.run(call.split(' '))
+    assert out.returncode == 0
+
+def test_gwas_bgenix():
+    call = f'singularity run {pth} bgenix -help'
     out = subprocess.run(call.split(' '))
     assert out.returncode == 0